study_notes_app/requirements.txt
# Core GUI
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0

# AI
anthropic>=0.40.0

# PDF processing
PyMuPDF>=1.23.0

# DOCX processing
python-docx>=1.1.0

# PPTX processing
python-pptx>=0.6.23

# OCR (handwriting)
easyocr>=1.7.0

# Video transcription
openai-whisper>=20231117
ffmpeg-python>=0.2.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Markdown rendering (server-side fallback)
markdown>=3.5.0
pymdown-extensions>=10.0

# Build
pyinstaller>=6.0.0
